Add a break to avoid a TP Autosolve case
Module returns an IndexOutOfRange Exception if less than three stops are selected. This is intended, more or less, but having TP autosolve and having it return an exception is NOT intended. As such, I'm adding a break into the for statement so that it'll skip straight to the count check and return null there.